makes CZISlide optional
